Move stories into own folder
