+++ client/src/stories/Header.stories.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Header } from "./Header";
+import { Header } from "../components/Header";
 
 export default {
   title: "Component/Header",
