+++ client/src/stories/Navbar.stories.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Navbar } from "./Navbar";
+import { Navbar } from "../components/Navbar";
 
 export default {
   title: "Component/Navbar",
