+++ client/src/stories/Searchbar.stories.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Searchbar } from "./Searchbar";
+import { Searchbar } from "../components/Searchbar";
 
 export default {
   title: "Component/Searchbar",
